packages/agents/requirements.txt
supabase==2.30.0
groq==1.2.0
ollama==0.6.2
requests==2.34.0
beautifulsoup4==4.14.3
python-dotenv==1.2.2
